post request project and updated project service and project dao

+++ src/main/java/org/soniakbew/models/ProjectProperties.java
@@ -35,7 +35,7 @@ public class ProjectProperties {
         this.finishDate = finishDate;
     }
 
-    public double getCommissionRate() {
+    public float getCommissionRate() {
         return commissionRate;
     }
